[tps] 시스템 관리 > 워터마크 관리 > 워터마크 조회 API 생성

+++ moka-core/moka-core-tps/src/main/java/jmnet/moka/core/tps/mvc/watermark/entity/Watermark.java
@@ -27,7 +27,7 @@ public class Watermark extends jmnet.moka.core.tps.common.entity.BaseAudit imple
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "SEQ_NO", nullable = false)
-    private Integer seqNo;
+    private Long seqNo;
 
     /**
      * 출처
